A Shitload of Untested Code.
Yay untested code and incomplete code! It's updating the Settings
framework so that it can be dynamically populated based on the device
you select and the options for that device that you select. I can also
gauruntee that it won't work as of this CL, go back a few CL's.

I DO WHAT I WANT.